Refactor UI structure by moving dialog imports to a dedicated module and changing directory of UI files
