+++ ui/__init__.py
@@ -1,0 +1,3 @@
+from .custom_widgets import ComplexMenu, MenuTooltipFilter
+from .settings_dialog import SettingsDialog
+from .epsg_dialog import EpsgDialog
