+++ ui/design_files/settings_dialog.ui
